+++ compiler-core/src/exhaustiveness.rs
@@ -221,6 +221,7 @@ pub struct Diagnostics {
 pub struct Match {
     pub tree: Decision,
     pub diagnostics: Diagnostics,
+    pub subject_variables: Vec<Variable>,
 }
 
 impl Match {
@@ -291,6 +292,7 @@ impl<'a> Compiler<'a> {
         Match {
             tree,
             diagnostics: self.diagnostics,
+            subject_variables: self.subject_variables,
         }
     }